rename paths
