+++ stories/BarChart.story.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { storiesOf } from '@kadira/storybook';
-import { BarChart } from '../components';
+import { BarChart } from '../src';
 
 const displayName = BarChart.displayName || 'BarChart';
 const title = 'Simple usage';
